[FEAT] 진동 On/Off 기능 추가
- SettingCanvas 수정: 진동 버튼 추가
- 버튼 클릭 시 버튼 이미지 변경 (진동버튼, 테마버튼)
- 아이콘 에셋 추가

+++ Assets/Scripts/Vibration/VibrationEditor.cs
@@ -9,7 +9,7 @@ public class VibrationEditor : VibrationInstance
         return false;
     }
 
-    public override void Vibrate(VibrationType vibrationType, float intensity)
+    public override void Vibrate(VibrationType vibrationType, int intensity)
     {
         throw new NotSupportedException();
     }
